requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
allure-pytest==2.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest-cov==4.1.0
